requirements.txt
watchdog
python-dotenv>=1.0.0
colorama>=0.4.6
requests
psutil>=5.8.0
pywin32; platform_system == "Windows"
pyinstaller 
DrissionPage>=4.0.0
selenium
webdriver_manager
